+++ cookbooks/base_setup/attributes/default.rb
@@ -1,0 +1,3 @@
+override['mysql']['server_root_password'] = 'rootpass'
+override['mysql']['server_repl_password'] = nil
+override['mysql']['server_debian_password'] = nil