+++ cookbooks/base_setup/recipes/default.rb
@@ -6,3 +6,19 @@
 #
 # All rights reserved - Do Not Redistribute
 #
+
+include_recipe "apache2"
+include_recipe "mysql::server"
+include_recipe "mysql::client"
+include_recipe "php"
+include_recipe "php::module_mysql"
+include_recipe "apache2::mod_php5"
+
+# Install packages
+%w{ mc }.each do |a_package|
+  package a_package
+end
+
+apache_site "default" do 
+  enable true
+end