+++ cookbooks/base_setup/CHANGELOG.md
@@ -5,7 +5,7 @@ This file is used to list changes made in each version of the base_setup cookboo
 
 0.1.0
 -----
-- [your_name] - Initial release of base_setup
+- Alexandr Lunyov - Initial release of base_setup
 
 - - -
 Check the [Markdown Syntax Guide](http://daringfireball.net/projects/markdown/syntax) for help with Markdown.